feat(FEV-1186): address PR comments
