+++ src/components/quiz-review/question-list-review/index.ts
@@ -1,0 +1,1 @@
+export * from './question-list-review';
