+++ src/components/quiz-review/question-list-review/question-list-review.scss
@@ -1,4 +1,4 @@
-@import '../../theme.scss';
+@import '../../../theme.scss';
 
 .quizReviewWrapper {
   display: flex;
